rename seedToNode to seedToRoot to fit in with the rest of the naming
