+++ index.js
@@ -22,7 +22,7 @@ import unspentOutputsToInputs from './src/unspentOutputsToInputs.js'
 import inputsToUnsignedTransaction from './src/inputsToUnsignedTransaction.js'
 import masterRootToAccountPrivKey from './src/masterRootToAccountPrivKey.js'
 import masterRootToAccountPubKey from './src/masterRootToAccountPubKey.js'
-import seedToNode from './src/seedToNode.js'
+import seedToRoot from './src/seedToRoot.js'
 
 export default class Elrn {
     constructor(options) {
@@ -41,8 +41,8 @@ export default class Elrn {
     seedToAddress (seed, derivePath, network) {
         return seedToAddress.call(this.options, seed, derivePath, network);
     }
-    seedToNode(seed, coinExt){
-        return seedToNode.call(this.options, seed ,coinExt);
+    seedToRoot(seed, coinExt){
+        return seedToRoot.call(this.options, seed ,coinExt);
     }
     seedToPublicKey (seed, derivePath, network) {
         return seedToPublicKey.call(this.options, seed, derivePath, network);
